refactor(runtime): collapse transport pool into pooled transport
